// frontend/src/types/herramientas.ts
export interface HerramientaInformatica {
    id?: number;
    nombre: string;
    descripcion?: string;
    funcionalidad?: string;
    responsable?: string;
    departamento?: string;
    fecha_creacion?: string;
    ultima_actualizacion?: string;
    estado: string;
    version?: string;
    ubicacion_archivo?: string;
    fallas_comunes?: string;
    fuentes?: string;
    observaciones?: string;
    ecosistema?: string;
    sys_fecha_creacion?: string;
    sys_ultima_modificacion?: string;
}